test de la fonction de capture de temps

+++ SoundGen-V1/src/pjl/appli/Grain.java
@@ -1,26 +1,26 @@
 package pjl.appli;
 
 public class Grain {
-	float frequence;
-	float instant_emission;
-	float duree;
+	double frequence;
+	double instant_emission;
+	double duree;
 
-	public Grain(float F,float dureeT,float instantT){
+	public Grain(double F,double dureeT,double instantT){
 		frequence = F;
 		duree=dureeT;
 		instant_emission=instantT;
 		}
 	
-	public  float[][] getAmplitude(float t,int n){
+	public  double[][] getAmplitude(double t,int n){
 		
-		int Nbre_Ech = (int)((this.duree)*8000);
+		double Nbre_Ech = (double)((this.duree)*8000);
 		
-		float echantillons[][]= new float[2][n];
+		double echantillons[][]= new double[2][n];
 		
 		for (int i = 0;i<n;i++){
-			echantillons[1][i]=t+(float)i/8000;
+			echantillons[1][i]=t+(double)i/8000;
 			float w = (float) Math.exp(-0.5*(10*((i-Nbre_Ech/2)/(Nbre_Ech/2))));
-			echantillons[2][i]=(float) (w*Math.cos(2*Math.PI*this.frequence*(echantillons[1][i]-this.instant_emission)));
+			echantillons[2][i]=(w*Math.cos(2*Math.PI*this.frequence*(echantillons[1][i]-this.instant_emission)));
 			}
 		
 		return echantillons;
